Further reduce code duplication
