+++ tests/Unit/CodeParser/FullyQualifiedClassNameResolverTest.php
@@ -2,6 +2,7 @@
 
 namespace JonasPardon\LaravelEventVisualizer\Tests\Unit\CodeParser;
 
+use Exception;
 use JonasPardon\LaravelEventVisualizer\Services\CodeParser;
 use JonasPardon\LaravelEventVisualizer\Tests\TestCase;
 
@@ -72,5 +73,34 @@ final class FullyQualifiedClassNameResolverTest extends TestCase
         $resolvedClass = $codeParser->getFullyQualifiedClassName('SomeEventAlias');
 
         $this->assertEquals('App\Events\SomeEvent', $resolvedClass);
+    }
+
+    /** @test */
+    public function it_throws_when_multiple_imports_are_defined_on_one_line(): void
+    {
+        $this->expectException(Exception::class);
+        $this->expectExceptionMessage('Multiple imports in one line not supported for now');
+
+        $codeParser = new CodeParser(
+            <<<'CODE'
+                <?php declare(strict_types=1);
+                
+                use \Event, \Bus;
+                
+                final class ClassName
+                {
+                    public function __construct()
+                    {
+                    }
+                
+                    public function classMethod(): void
+                    {
+                        Event::dispatch();
+                    }
+                }
+                CODE
+        );
+
+        $codeParser->getFullyQualifiedClassName('Event');
     }
 }
