+++ tests/Unit/CodeParser/MethodCallsParsingTest.php
@@ -32,40 +32,6 @@ final class MethodCallsParsingTest extends TestCase
         $this->assertEquals($expectedMethodCalls, $methodCalls);
     }
 
-    /** @test */
-    public function it_throws_when_multiple_imports_are_defined_on_one_line(): void
-    {
-        $this->markTestIncomplete('Not implemented yet');
-
-        $this->expectException(Exception::class);
-        $this->expectExceptionMessage('Multiple imports in one line not supported for now');
-
-        $codeParser = new CodeParser(
-            <<<'CODE'
-                <?php declare(strict_types=1);
-                
-                use \Event, \Bus;
-                
-                final class ClassName
-                {
-                    public function __construct()
-                    {
-                    }
-                
-                    public function classMethod(): void
-                    {
-                        Event::dispatch();
-                    }
-                }
-                CODE
-        );
-
-        $codeParser->getStaticCalls(
-            subjectClass: 'Illuminate\Support\Facades\Event',
-            methodName: 'dispatch',
-        );
-    }
-
     public function providesCodeSamples(): array
     {
         return [
